+++ Sources/PeripheryKit/Formatters/OutputFormatter.swift
@@ -105,8 +105,8 @@ public extension OutputFormat {
             return JsonFormatter.self
         case .checkstyle:
             return CheckstyleFormatter.self
-        case .actions:
-            return ActionsFormatter.self
+        case .githubActions:
+            return GitHubActionsFormatter.self
         }
     }
 }