+++ Sources/Shared/OutputFormat.swift
@@ -6,7 +6,7 @@ public enum OutputFormat: String, CaseIterable {
     case json
     case checkstyle
     case codeclimate
-    case actions
+    case githubActions = "github-actions"
 
     public static let `default` = OutputFormat.xcode